+++ src/helpers/healthBar.js
