+++ src/helpers/helper.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-unresolved */
 import Phaser from 'phaser';
 
-const score = require('./Scenes/Battle');
+const score = require('../scenes/battle');
 
 const scoreTextUpdate = (el) => {
   setInterval(() => {
