+++ src/scenes/titleScene.js
@@ -2,8 +2,8 @@
 /* eslint-disable import/no-unresolved */
 
 import Phaser from 'phaser';
-import config from '../Config/config';
-import Button from '../Objects/Button';
+import config from '../config/config';
+import Button from '../helpers/button';
 
 export default class TitleScene extends Phaser.Scene {
   constructor() {
